video_manager_api/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
redis==4.5.1
google-cloud-pubsub==2.21.1
python-dotenv==0.21.0
python-multipart==0.0.20

